requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
psycopg[binary]==3.2.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.10.4
anthropic>=0.77.0
python-multipart>=0.0.9
